+++ src/main/java/com/jfcm/manda/bookingmanagerapi/exception/CustomExceptionHandler.java
@@ -70,6 +70,20 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
     return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
   }
 
+  @ExceptionHandler(MultipleBookingException.class)
+  public ResponseEntity<ErrorResponse> handlesMultipleBookingsException(MultipleBookingException ex, WebRequest request) {
+
+    var error = ErrorResponse.builder()
+        .timestamp(dateNow)
+        .status(HttpStatus.FORBIDDEN.value())
+        .responsecode(Constants.TRANSACTION_FAILED)
+        .message(ex.getMessage())
+        .path(request.getDescription(false))
+        .build();
+
+    return new ResponseEntity<>(error, HttpStatus.FORBIDDEN);
+  }
+
 //  @Override
 //  public ResponseEntity<Object> handleMethodArgumentNotValid(
 //      DataAlreadyExistException ex, HttpHeaders headers,