+++ src/main/java/com/jfcm/manda/bookingmanagerapi/repository/ReservationRepository.java
@@ -7,10 +7,13 @@
  */
 package com.jfcm.manda.bookingmanagerapi.repository;
 
+import com.jfcm.manda.bookingmanagerapi.model.BookedEventsEntity;
 import com.jfcm.manda.bookingmanagerapi.model.ReservationEntity;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -21,4 +24,6 @@ public interface ReservationRepository extends JpaRepository<ReservationEntity, 
   List<ReservationEntity> findAll();
 
   void deleteById(String id);
+
+
 }