+++ src/test/java/com/jfcm/manda/bookingmanagerapi/resource/ClustersResourcesTest.java
@@ -1,7 +1,7 @@
 /**
  * {@link com.jfcm.manda.bookingmanagerapi.resource.ClustersResourcesTest}.java
- * Copyright © 2023 JFCM. All rights reserved. This software is the confidential and
- * proprietary information of JFCM Mandaluyong
+ * Copyright © 2023 JFCM. All rights reserved. This software is the confidential
+ * and proprietary information of JFCM Mandaluyong
  *
  * @author Ronald Cando
  */
@@ -21,19 +21,14 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jfcm.manda.bookingmanagerapi.model.ClusterGroupsEntity;
 import com.jfcm.manda.bookingmanagerapi.repository.ClustersRepository;
-import com.jfcm.manda.bookingmanagerapi.service.impl.GenerateUUIDService;
-import com.jfcm.manda.bookingmanagerapi.service.impl.LoggingService;
 import com.jfcm.manda.bookingmanagerapi.utils.TestUtils;
-import com.jfcm.manda.bookingmanagerapi.utils.Utilities;
 import jakarta.transaction.Transactional;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -49,9 +44,9 @@ import org.springframework.test.web.servlet.ResultMatcher;
 @Transactional
 class ClustersResourcesTest {
 
+  private final ObjectMapper mapper = new ObjectMapper();
   @InjectMocks
   private ClustersResources clustersResources;
-  private final ObjectMapper mapper = new ObjectMapper();
   @Autowired
   private MockMvc mockMvc;
   @Autowired
@@ -184,7 +179,7 @@ class ClustersResourcesTest {
       resultMatcher = status().isOk();
     } else {
       //clustersData = TestUtils.readFileValue(mapper,
-        //  "json/test-data/single-cluster-data-for-update.json", ClusterGroupsEntity.class);
+      //  "json/test-data/single-cluster-data-for-update.json", ClusterGroupsEntity.class);
       clusterId = "cluster-000";
       transactionMessage = "doesn't exist";
       tranCode = "TRN-001";